requirements.txt
Django==1.8.5
django-filter==0.11.0
django-health-check==1.0.2
djangorestframework==3.2.4
djangorestframework-filters==0.5.0
djangorestframework-jwt==1.7.2
Pygments==2.0.2
PyJWT==1.4.0
requests==2.8.1
